remove static_entitlements file
